rename operations file to query_operations
