+++ marketplace/plugins/s3/lib/query_operations.ts
